Fixup for a7ec6fc6ac62 (Fixed: Bug 35199 - mail order incorrect)
The commit broke the tests, because of some edge-cases not working any
longer. Adjusted the mock implementation of fetch to work for the
general case and extendend the test a little (no functional/external
changes).

Now the implementation works as expected, again. Loading one element more
than the actual limit.