fix namespace-based loading
in fact this is a cleanup of the precore.js module that is generated
manually to speed up loading times by decreasing the amount of requests
needed to "boot" into appsuite.

The file now contains a proper module definition with dependencies and
the definition of the pre loaded modules is deferred until all those
dependencies are met.

This solution is less invasive than c8429ad0de7ec822217421410fd since
no behavioural changes to our libraries are required.