fix loading behaviour of precore.js
if the user had been signed in before module definition function of
precore.js had been run, the event was already triggered and so the
loaded modules never got defined. Now both cases are handled, if the
user is logged in, the config is loaded directly. If this is not the
case, config is loaded once the user is logged in.