Fixed: Bug 37680 - [L3] Catalan translation does not work on Appsuite
manually update the po file from pootle, because it had not been synced
for some time, now. Since ca_ES is not part of our translation process,
but provided by the community, the translation did not make it into the
release. This is a manual sync of the file from l10n.open-xchange.com.