Fixed: Bug 35944 - Attaching a file after removed it in mail compose
This is actually a browser issue, but there seems to be a workaround.
Problem is, no change event is triggered, if the same file is selected
again in the OS File dialog. By reseting the value of the input field
every time, this event will be triggered.

Tested in the major browsers (Chromium, Firefox and IE)