fix rhino version of update-themes
fix is done by importing references with their full path in top of each
less file and on the other hand switching to the node implementation of
Parser.fileLoader (see last commit)

This code is based on the implementation used by assemble-less project,
which we use via grunt. Referenced source files are imported with their
full path, instead of only the filename read from the options. This
allows less to pick up the base path if the imported file itself
contains imports.