Fixed: Bug 37807 - Document not editable although permissions allow editing
the created_by field is needed to check permissions correctly in this
case. This is a very greedy approach to fix this bug. It might be more
efficient to create get requests for a selection in order to retrieve
the information, on the other hand, this seems to be a cheap operation
and should not have a huge impact. Until this proves to be slow, the
simple solution might be the better one.