Fixed: Bug 38373 - [L3] screen gets blurred when using search on iOS
there has been a little work-around in place to handle menu painting in
a custom way when using menus from within the search application. Since
search has been through very large changes during the versions, it's
okay to extend the work-around to also work on iOS devices.

For later versions of search this is not needed, because the code has
been rewritten and the work-around seems not to be in use any longer.