if appserver https options link to files, read those
in the config file, developers might actually provide a path to a file
instead of the actual data. Test if this is the case and then import the
data from that file is a sane thing to do. The error message is not very
informative, when passing in a filename instead of the data to connect.