Fixed: Bug 40630 - [PC,Win10] Upload folder via drag&drop causes unknown error
Try to detect, if a folder is uploaded. This is not 100% reliable. But the newest versions of Chrome, Safari and Firefox will inform the user, that directories cannot be uploaded. Since the internet explorer does not throw a drop event when a folder is dropped, this handling cannot be applied to the internet explorer.