Fixed: Bug 40568 - Mobile view: missing breadcrump in drive
Instead of the breadcrumb, drive does now provide a 'move to parent folder' icon. With this icon you can always move to parent and you are able to navigate threw the folders without changing to the root folder. If a folder has the root folder as parent, the icon will switch to the folderview.