Fixed: Bug 41147 - Themes: leaking global state from pre-login into login procedure
once we are not in login mode anymore, the theme should be reset. If
this is not the case, less! plugin will load any CSS files with the
prior theme (which is the login theme, which is most likely not a theme
in the sense we define it). By resetting the global state, the less!
plugin will not load any CSS files, until theming is initialized during
the boot process.