Fixed: Bug 41276 - [Nexus 5, 5.1] - calendar - While creating event default date is always today
Set default date to the refdate of the app when not in listview. In dayview, the inital date in the edit dialog will always be at the selected day. In month and week view, the date is at least in the visible range. Depends on which view was selected before. The inital date for week and month view could be optimized but I am not sure that this will not produce any sideeffects.