Handle id changes of files
If the id of a file changes, the id and cid of the file will be updated and therefore any changes to the collections will directly go to the model.
This has the effect that views which use the model where the id/cid changed will get further updates of that model.

(cherry picked from commit 14430a6a6df4112c6612fe6d0a7e28e58758b380)