Fixed: Bug 42634 - mail standalone view crops content
Set images in mails to a max-width of 100% when the mail is considered simple.
A mail is not simple, if it contains tables. I looked at several common mails with a lot of markup and some simple mails and simple mails usually don't contain tables and mails with lot of markup often contain tables.