Fixed: Bug 41709 - [L3] Portal mail widget emtpy
Reload mail data when collection is set to expired.
This fix does also include the handling, when mails are moved to the folder which is displayed in the portal. The target folder is then updated.

(cherry picked from commit 75e8aca9dfc42b2bc54fa1b428d78960aed5d9a7)

 Conflicts:
	ui/apps/plugins/portal/mail/register.js