Better fix of regression: Bug Bug 45364 - Content spoofing at login page
adding the type to the message in some way is indeed usefull for
debugging purposes. Bug reports mainly contain the message and it might
help to find the root cause just by seeing the login method used. In
order to not allow for injection of arbitrary content, limit the type to
the first word that looks like it could be a method.