Fixup for Bug 45798 - [L3] Quick Reply not correctly translated
Fixes commit f55add31858d50b66e05c0f371b27f79bd935725, which broke the
build of language files.

Technically this is not a plural decision at all, the string to be used
just happens to depend on the number of recipients. Its just, that the
action is named slightly different if only one participant is involved.
There should be no difference in any language for more than one
participant. If there really is such a use-case, npgettext must be used
and a context added, because "Reply" as a msgid is already present from
other parts of the software and can therefore not be used a second time.