Fixed: Bug 45296 - [L3] some appointments not shown in month view
Root cause: calculation of week borders was done by hand, and did not
consider all timezone data
Solution: use moment to calculate everything

This is a collection of fixes from develop branch for the 7.8.2 release.
Since the calendar module did not change much, in order to simplify
future backports for the 7.8.1 branch, basically synchronize the code.
(This version of moment.js doesn't support the ::isSameOrBefore method,
so this needed to be implemented differently)