remove commented code
This was commented with the intend to never activate it again, because we discussed the reason this code has
been there in the first place is not valid any longer. If the code is really needed, we can always get it
back from the git history. That's what SCM is for.