allow custom sort index for facets
currently there is only a very limited way to customize the order of facets.
This allows the autocomplete middleware to specify a custom index, which is then used
to sort the list of facets shown to the user.
Of course, this can now be customized using the autocomplete extension point by injecting a
custom index into the result set.

(cherry picked from commit 1726da4becbee25ee46fbeebb33668cf3af85d9a)