Implement OXUI-188: Separate JS code loading for advertisement plugins
This was a follow up Story of Bug 46159 - Advertisement framework
with active AdBlock blocks loading of UI - which is now considered
fixed.

It is still possible to load all advertising plugins manually without
using our "code concatenating" feature, but it can also be implemented
by adding some code to handle the error produced during the request.
This solution is even more general and should handle other loading
problems as well. Not only ad blocker related ones.

The solution is pretty simple: if a request fails, there is now a
fallback to separately load the requested modules. A dummy module is
registered for single modules which fail to load. This minimizes the
amount of code missing, if only one module can not be loaded.