Fixed: Bug 48986 - Wide/long mails not shown properly, space wasted, layout broken
As there was no general fix for this we introduced a new option and some logic
to autmatically choose the right size for wide emails. When opening an email
in the single page mode (doubleclick) the user can choose to toggle the view's
width by pressing a new button on the top right in the mail detail view. If the mail
body width exceeds a threshold of 800px the wide-view mode will be toggled automatically.