Fixup for commit 8f2d4a1cc4e22b7cb8a54598b9d5211ddca3ccf2
Fixed Bug 49588 - Strange behavior of add storage function

-----

this simplifies the code quite a lot. Adding an already existing account
to a collection is a noop, no need to check if the account actually is new.
So the fix should rather be to only add the account, if it is saved to
successfully the backend.