implement packaging stage for ci pipeline
use environments, so the packages will be removed automatically once the
environment is stopped. In the packaging stage, source packages will be
created and uploaded to OBS.

only build on a few branches those branches will never collide with any
jenkins build, because they would fail on OBS, because of slashes in branch
names.