unify visual appearance of service provider list
Since the implementation of "add OAuth based mail accounts", there has
been a general implementation of a selectable service provider list.
Visually, this list was very different from what was implemented for
file storage accounts. The general implementation now looks and works
the same, as the file storage implementation, providing a base to
simplify and unify code wherever a OAuth service provider list is
needed. Of course, everything is based on Backbone, now.