port add-storage-account to new internal OAuth API
this moves a lot of the complexity into the OAuth module.

TODO: currently it's not possible to create a second filestorage account
for the same service. Middleware needs to "fix" that, since it's
possible to have multiple mail accounts for the same service, it should
also be possible to have multiple storage accounts per service.