OXUI-277: Generalize service provider selection list of file storage accounts
do not disable selection buttons after selecting the provider. There
seems to be no reliable way to get notified if the popup window is
closed, again. The unload event is triggered, as soon as the page
navigates to the provider for authorization of the scope, and can not be
registered again, due to CORS restrictions.