cleanup OAuth icon handling
icons are now determined by serviceId in combination with a few
exceptions for special services with multiple logos, one per service.
The decision is handled completely in CSS/LESS, using the (service-)id
as a base. This should be extensible and can be easily customized just
by providing a theme pointing to custom images.