add a limit for requests sent to apps/load servlet
collecting all known modules might result in very long urls which might
be limited on the server. By default, an apache server limits the
request line to 8190 bytes. Since we know the length, we can split large
requests into multiple ones, none of which will exceed the limit. In
order to allow administrators to customise this value according to the
specific environment, this value can be configured in the serverConfig
attribute which is part of the rampup call.