Fixed: Bug 51607 - Mail standard folders are not recognized as standard folders in case the folder is a subfolder
List all default folder types and filter for available folders. Those folders are added as default folders below the inbox.
All sub-subfolders which are used as default folders will appear twice in the folder tree (below inbox and in the folder tree). This cannot be solved easily, because then, we cannot rely on server side subfolder flags (for example, if such a sub-subfolder is the only subfolder of it's parent).