Fixed: Bug 52540 - Can't create/edit email signatures
Root cause: missing global tinyMCE code when rendering the editor
Solution: require the needed code before actually initialising the
    editor
Resolution state: fix

initially there has been a hard dependency on the tinymce file, but it
has been removed for an unknown reason. This changes defers the loading
of the file until it really is needed. If there are still timing issues
with loading the code in some situations, those should be fixed instead
of removing the dependency, again.