replace appserver middleware modules with official ui middleware
instead of maintaining two different code base, we are currently working
on an official nodejs implementation of a few servlets which used to be
re-implemented in appserver. To have a wide coverage of the production
code, the new nodejs middlerware components should also be used for
local development as much as possible.