fix e2e tests
change e-mail address to match server setup, which does not use QA
mailserver any longer, because this should now be more stable.

Also mail-detail view is now ported to use iframes instead of shadow
DOM. Unfortunately it's not easy to switch the frame context, because
there is no frame id. So for now, mail body is not really checked.