Fixed: Bug 56219 - Missing translation on login page
since the file which contains the module does not import the gettext
module being used, the strings were not extracted into the pot file.
Since boot is a little special, strings exclusively used during the boot
process need to be listed in a dummy module for the translation process
to pick those up.

In the future it should be possible to define the local gt variable and
it's related namespace/translation module via some kind of metadata
mechanism, to remove the need for dummy modules and therefor to prevent
such errors from happening.