Fixed: Bug 56406 - Content of mails added "to Portal" is being executed as script code
Root cause: code to "unescape" mail content actually executed the code;
    This has to do with the format in which plain text mails arrive.
    Nowadays, the middleware sends back the original mail body, while initially
    all HTML entities had been escaped. In order to not insert escaped version of
    the mail body, but the original one, the mail body needed to be unescaped.
    This only works under the assumption that there is no HTML code within the
    text to unescape. While this has been true for earlier versions, it is not
    the case any longer.

Solution: remove "HTML unescape" code, since it's not needed any longer, but
    introduces an XSS instead.

Unfortunately this can not really be tested using unit tests. In order
to simplify backporting, tests are not added with this fix, but will be
provided in a separate commit. Since we now have e2e tests, it is
possible to effectively test for this defect, so a regression is
unlikely to happen.