Fixed the "u is not a function" problem with update-themes when using Rhino
Root cause:
A function declared in a compound statement (as opposed to the top level
of the parent function). Rhino does not seem to lift such function
declarations to the top of the block.
The problem started occuring when the uglifier suddenly started moving
the declaration to the bottom of the block instead of the top.

Solution:
Use a function expression instead of a function declaration.