remove workaround for unit test environment
this solution seems a little more clean to me, it doesn't require a
running mail app, which used to be a little fragile in the past. The
problem with the last app being closed and the default app (which is
mail, if not configured) being opened is now mitigated by just setting
the autostart app to none.