OXUI-489: Show "department" field for all contact search results
This is a first drop to implement such a feature. The idea is to extend
the core extension point to "draw" a person (io.ox/core/person) to show
the department if the setting is active, the contact is from the GAB and
has a department set. Since this is all extension based, it can be
customized as usual.