Fixed: Bug 57086 - All Day appointment is updated wrong
Root cause: regression of the fix for Bug 57061 in save button handler
Solution: use util function to determine "allDay" attribute from a model

This is most likely not the fix for the initial bug report, which I
guess has been fixed with Bug 57061. At least the use-case works for me,
now. However, manually changing the "length" of the appointment was
still off-by-one.

This fix also brings an e2e test for exactly this scenario.