Sort apps after index attribute in manifest.json files
Until now, the setting io.ox/core//apps/order had a default, enforcing a
certain order of the apps - if not configured otherwise. This rendered
the index attribute of manifest files more or less useless. This change
reverts this behaviour, making the manifest files the default, allowing
admins to override using the setting and shorten the list by maintaining
a blacklist.