add error handling for oauth init action in reauthorize case
reauthorization needs to be done with the same identity at the oauth
provider level. If the identity for this workflow does not match the
oauth account details, an error will be thrown. In this case, we only
show the error message and abort the process.

May be, a dialog should be shown, explaining the situation.
A 'retry' button might be provided to resolve the situation more easily,
without starting all over, again.