Fixed: Bug 58019 - [L3] Mailbox unable to forward multiple emails
Root cause: workaround to fetch mail before composing doesn't work with
    multiple source mails involved
Solution: check for multiple mails and skip workaround in this case

In order to limit possible side-effects of this for systems that don't
use this workaround, the check for the setting to activate the
workaround is now done first.

I tested a little more, but it's very hard to test all possible
use-cases and the workaround still might have introduced some unwanted
behaviour (like in this case).