remove grunt-modernizr as devDependency
this currently breaks yarn install and is not needed any longer. We will
probably remove modernizr completely for 7.10.1 release. For the time
being the lib is part of the repo anyway and we don't need to update it
currently. If there should be the need to update, it can be easily
installed semi-automatically, again. Really no need to track this as a
(dev-) dependency