explicitly set versions for potentially unsafe dependencies
unless we control the releases of the libraries, it might always be
possible, dependencies break the build if there are incompatible updates
in minor versions. To make sure, npm install produces an output as close
as possible to yarn (using a lock file), at least for our builds, we
explicitly want to specify the desired versions.