[dynamic theme] remove optional request library for less
I don't think, this library is ever used, it's optional and only
required to support URLs in import statements. Doesn't make much sense
to me in browser environments. This reduces the maximum path length
about 80 characters. This should result in less problems when cloning
the repo on windows (which has a path length limit of 256 characters)