Fixed: Bug 58882 - Add storage account dialog doesn't evaluate canAdd of external drive
Root cause: canAdd of keychainAPI not evaluated when drawing available
    services list
Solution: also evaluate canAdd of keychainAPI, if available

this is only relevant for older plugins, which implement keychain API to
be checked wether an account can be created for a service or not. This
is also not relevant for 7.10.0 release, but only for 7.8.4.