Fixed: Wrong used cached models in list request
Root cause: When performing a list request, it was only checked if the model was present and not if the model has all fields before loading it. That caused models not to be fully loaded by a list request when the model has already been loaded by an all request.
Solution: Add an additional check and also request models which do not have all data.