Fixed: Bug 59226 - Notification window displays out of OX screen if Ads are installed
Root cause: fixed positioning is least flexible in terms of
    customizations
Solution: remove fixed positioning and calculate max height in code

visually looks the same, but CSS is simpler now and this works fine if
layout is customized. The drawback is, that we need to calculate the
max-height depending on height of the actual content. We already do
something similar for other elements in the topbar, so this should be
okay.