remove mfa error handling from http module
this also removes the "http.disconnect" code, again. Currently we are
not sure, wether this still is needed and we assume, everything will
work without such code. The error handling is now done in a separate
module, and everything can be controlled from there. The current
implementation uses online/offline state to communicate about
availability of the MW API. But this still needs some more testing and
work, until it can be used in production.