make calendar tests more stable on weekends
this will make sure, we are always creating a new appointment at the
beginning of the next week. If the test runs at a weekend, it will be
broken, because the appointment will be created in workweek view and
switching to "day" will select the wrong day. This now works fine.