Fixed: Bug 61481 - mail wrongly removed from grid
this is the second attempt to fix this. The first version did not work,
because of using the API in a wrong way. The bug itself disappeared by
accident, though, causing a regression when deleting mails due to a
javascript error. The new test now tries to detect this by adding more
use cases for deleting mails in the list view.

Root cause: item removed from listview if the thread defining mail is
    deleted.
Solution: reload if view is threaded and we most likely have deleted a
    relevant mail