OXUI-555: As an OAuth user I want to easily rename folders connected to accounts
Since it should now be possible to rename external Drive folders, the
possibility to do so in the setings pane for accounts is not needed any
longer. The extension to render the "name" field for filestorage
accounts is therefore removed. Another check in the DetailView renderer
will only render the edit button, if there actually is a settings/detail
extension defined for the account.