Fixed: Bug 62263 - [L3] Add mail account on mobile: buttons hide text field for input
Removes button 'manual' cause is it used as fallback in case autodiscover
fails and should not be handled as a separate option for ux reasons. As a
positive side effect of this change the bug get's fixed by the fact that the input
field is visible again on smartphone.

(cherry picked from commit 49818695eabe216701e67fb3ac75bd65fe09d57e)