Fixed: Bug 62935 - drivemail: if sending drivemail with expiry_date set fails or is aborted, resending puts expiration far in the future
Root cause: Mail model contains expiry_date as a delta in ms. Before send, this delta is added to the current timestamp. If send is aborted (due to whatever reason), the model will contain absolute timestamp, not a relative timespan and another call for send will compute a absolute timestamp again.
Solution: Make sure, that while sending, only a deep copy of the value is used to compute the absolute timestamp