switch to allure for e2e reporting
this seems to be a more stable approach for reporting codecept results.
Allure is compatible with xUnit XML files, so it can be used to report to GitLab
and Jenkins (and other tools, of course).
It also allows to use the allure toolchain to generate nice HTML reports.

On the other hand, allure seems to be much more complicated and we have to rely
on external people to maintain the software (probably forever)