Fixed: Bug 63234 - Inline image is inserted in the wrong compose window
Root cause: A lot of references considered the tinyMCE.activeEditor, which refers to the last active editor. If multiple windows are opened, this can lead to problems if actions are invoked (like image drop), which do not require the activeness of the editor window.
Solution: Always refer to the the editor, where the code is actually invoked