Fixed: Bug 60740 - Concept of frontend files in middleware and Apache MUST be documented
Root cause: missing/outdated documentation how theming is supposed to work
Solution: add more clear documentation how theming and code loading in general
    is supposed to work.

This change does not include explicit documentation for "touch-appsuite" tool,
which is part of operator tool belt when running App Suite systems.
The tool ships it's own documentation as part of a "--help" option.
It's not really possible to integrate documentation for operator tools,
because there currently is no section for that in our documentation and
I'm unsure we do want to maintain such a thing.