Fixed: Bug 63441 - [iOS] Keyboard appears after every click after editing a mail
Root cause: When having the cursor (and of course the focus) inside an iframe, a click on a button will not trigger a blur event inside the iframe. Therefore, the iframe stays selected and will try to regain the focus on every subsequent focus change. That leads to the opening of the keyboard for every other click which is not on a focusable item (means a link).
Solution: Change button to a link. Then, the iOS focus magic seems to work.