Fixed: Bug 63498 - Odd scrolling in mail compose iFrame on Safari
Root cause: When scrolling with the touchpad and the size of the mail was large enough, the iframe captured rubberband scrolling while the finger was on the touchpad. When the hand was released, the scroll was forwarded to the outer container and scrolled correctly.
Solution: Prevent scrolling of html stops rubberband on iframe