Fixed: Bug 60340 - [A11y][DEQUE:230733] Contacts/Calendar - Copy/Print Modals - Keyboard Trap (partial)
The iframe is now aria-hidden as it serves only presentational purposes, as setting it as role="presentation" would still annoy screen reader users.
Also added a fieldset so that they now are read properly as part of a group with a reference to the title.

(cherry picked from commit f11b79178dd370e5bf3b8508172f51361993770c)