Fixed: Bug 63902 - Unable to cancel sending mail
Root cause: Currently, the abort button is still displayed if the mail send progress is at 100%. But cancelation of mails with progress of 100% is not possible because the abort of the post-request might reach the server after the mail has already been send. Therefore, this is disabled but the abort-button was still show.
Solution: Hide abort button if send progress is at 100%.