Fixed: Bug 64163 - 'View', 'Download' and 'Save to Drive' in mail isn't working
Root cause: Rerendering of the attachment view caused the toolbar to be removed from the dom once and the events where not reapplied, since the toolbarview extends the disposable view.
Solution: Rerender the views as needed and reuse parts as needed. Additionally, make sure that no events are attached twice