Fixed: JS error on viewer close
Root cause: Close was invoked manually by the viewers mainview. This disposed the toolbarview which itself runs the click listeners afterwards. This leads to an error because a disposable view should not be used after it has been disposed
Solution: Process close event via extension point which is invoked by the toolbarview.