Fixed: Bug 65410 - [L3] calendar items in shared / public calendar are always using the claendar color for other users but the owner
Root cause: Appointment color was only considered if the user is the owner of an event such that the user can select the color for the whole public folder (see OXUI-560)
Solution: Appointment colors are considered for organizers and organizers_on_behalf.
Please note that this only applies to shared calendars. Public calendars will not change