backport packaging pipeline to remove nodejs dependency on OBS
since it doesn't seem possible to get a recent version of nodejs packaged
on RHEL6, we aim to backport the latest changes on how we do packaging to older
branches. This should help supporting those versions in the future and minimize
the effort needed to keep the build and packaging pipelines green.