Fixed: Bug 66556 - [L3] floating windows for mail compose can not be closed
Root cause: When opening a restorepoint, the id is incremented. But for objects from the jslobs, the object reference is still pointing to the object in the jslobs. Therefore, the id in the cache is also changed and the object with the old id cannot be found and deleted.
Solution: Work on a copy of the object to prevent to overwrite the id in the jslobs object.

(cherry picked from commit 323c6cc2ac2b3f9c852e29fa91ea2dea5889b2f2)