OXUI-717: Improve OIDC relogin flow
relogin handling has been refactored completely and is now realised using
the Stage pattern.
This allows easy customisation and the code should be more easy to follow.
There used to be two different implementations for dialogs, which are
now unified and the dialog is now really extensible (got assigned an
extension point).
The actual improvement for OIDC is, that an iframe is now used to regain
a valid session from the middleware.
If this succeeds, the user won't notice any interruption.