Fixed: Bug 67556 - appointment Search works not consistent
Root cause: Search tried to expire the actual collection which is bound to the apps collection loader whereas search has an own collection loader with a different type of collections. This is special for calendar, because search has it's own collection loading mechanism.
Solution: Expire the current collection of the search's collection loader