OXUI-717: oidc relogin flow improvements
until now, there has been a timeout to actually show
the session lost dialog in case the silent relogin
does not respond with a session after 10s.
Within the dialog, a new handler will be registered
still waiting for any response.
If there is a late response, the dialog will be closed
automatically.