Fixed: Bug 67628 - Preview pictures have wrong colors
 Chrome 77 and higher have a bug in the offscreenCanvas function
causing a red-blue shift for all images resized with the canvas.
This is the case for mail compose attachment previews and resized
images. As long as the bug is not fixed we disable the offscreenCanvas
which will make is slower on Chrome but prevent wrong colors

(cherry picked from commit 198f3206de2c2e09642ea07da8b139f2f8126718)