Fixed: Bug 66217 - viewOptions in settings!io.ox/files JSLob module might grow too large
Root cause: no garbage collection on viewOptions jslob data when removing folders
Solution: remove viewOptions data for folders being deleted

In case of an error while saving the viewOptions, the data will be reset
and only the last value will be stored. This should help all users that
have to much data stored with viewOptions.

Since the exception is due to the number of keys of the data object, this might
still be a problem if the user has 2500+ folders. In other words, currently there's a
hard limit of 2500 folders until users start running into this problem. It's not
possible to change this without comming up with an alternative way of storing
those options. However, I would not consider this to be in the scope of this
bug.