Bug 67805 - Mobile Text - Can't insert image
There is a query to for devices that are not a desktop browser. This query disables the action connected to the "Upload local image" button
This happens also in the mail compose dialog for attachments

Root cause: There is a query to detect desktop browsers. for mail compose there were no options related to the uploadButton set.
Solution: Removing the query that disables the upload button and adding options for the uploadButton