Fixed: Bug 68034 - iPhone 7, Safari - Drive: After setting sort order, I have to tap twice to open sort dropdown again
Root cause: In Safari (Mobile and Desktop), the document.activeElement does not refer to the clicked item while the delegate handler is invoked. Therefore, the focus cannot be restored to that element after the dropdown closes.
Solution: As the target element of the click event always has to be the element which should get the focus after the dropdown closes, we can just refer to that element.