Bug 66627 - Version list is not updated if in opened (even after refresh)
The version list in the detail pane of Drive is not updated if opened (even after refresh). One has to close and open the version list manually to see the latest versions. The number of versions is correctly updated.

Root cause: When the change:versions event is triggered, the file versions of the old version list are drawn
Solution: When the new change:versions is triggered update the versions of the file before drawing the version list in the details pane