Fixed: Bug 68051 - [A11y][Mild] Folder tree - Open folder view button toggle focus
Root cause: The focus stays on the button for opening or closing the folder view.
Solution: While opening the folder tree set the focus on the preselected element. If there is no preselected element, set the focus on the first item of the folder tree. While closing the folder view, set the focus on the first preselected element of the list view or last preselected element of the grid. If there is no preselected element, set the focus on the first element of the list view or grid. If there is no element to select at all, select the empty scroll pane. (concerned apps: mail, calendar, drive, tasks, contacts)