Bug 68106 - OX Documents uses openInSingleTab instead of separate browsertabs on ox.io
The tabHandling is deactivated because the capabilities are set in the context and not at the server config

Root cause: It is deactivated because the capabilities are set in the context and not at the server config. So the capabilities are not present at the boot process
Solution: activate the tabApi without office check. Only opening the new windows are checked with office capabilities