OXUI-769: Browsers might receive mixed up UI source versions
The middleware expects a version.txt file to be present in the App Suite root
directory (next to the apps/ and manifests/ folders". A new "version_txt" task
in grunt will take care that this file is present in the build/ directory and
therefor make it into the packages for the middleware to consume.

Also `apps/load` as well as `manifests?action=config` calls need a new “version”
parameter for the middleware to pick up the correct version.

A new job in the CI will make sure no packages are uploaded for pushed tags which
do not match the version specified in the `package.json` file.