Fixed default logo size, moved README.txt to README.md
The updated logo image in 7.10.0 does not have empty space at the
top and bottom anymore. Therefore, it needs an explicit width to scale
properly, otherwise it grows until it fills the entire 64px of available
height.

Therefore, the old default of `logoWidth=auto` does not work anymore
either. Replaced it with `60` (`px` is added automatically) like in the
core.