switch to build date as package version string for code loading
the current implementation relied on the version string in package.json
being updated for each update. This currently should be true but relies
on a semi-manual process to bump the version number for each release
(i.e. revision). According to the release team, in the future there
should be less updates of package.json, one for each stable release.
By just using the build date as an indicator for the version of a
package, we still workaround the problem of the middleware distributing
wrong sources for older ui versions and this even holds true for
snapshot builds.