Fixed: Bug 68689 - [L3] Script errors in IE and Firefox (2)
Root cause: On focus of an element, blur and focusin is triggered. That removed and added the 'has-focus' class, respectively. Changing classes of parent elements with styles based on that class will trigger a complete rerender of the children.
Solution: Debounce add and remove a bit, such that the has-focus class is only removed when really necessary. Prevents long running script when selecting.