Fixed: Bug 68689 - [L3] Script errors in IE and Firefox (5)
Root cause: When a new message arrived, the index of all messages changed which triggered a rerender of all list elements which could cause performance issues on large lists.
Solution: Omit index from head field to prevent changes on the head. Additionally, check if only the index of an element changed and do not rerender the element itself.