Fixed Bug 64173 - Tab -Popup blocker - open doument ( Upload local file):=> Popup blocker is shown
If a local file is uploaded in the open document dialog in a office app a popup blocker is triggered.

Root cause: There is a delay between user interaction and opening the new tab.
Solution: Don' open the uploaded file directly. It is selected in the dialog and the user must click on open.