improve touch-appsuite singlenode installations
for Ops, when running App Suite in larger clusters, it might help
to have a unified tool for cache busting installed on nodes.
Since touch-appsuite and the new tool for history management in the MW
need the exact same value for "timestamp", it makes sense to extend
touch-appsuite so that it can also be used as a minimal wrapper for the
new MW tool.

Since touch-appsuite is currently packaged with open-xchange-appsuite,
which has a dependency on httpd/apache, it only makes sense to use this
on singlenode installations. In theory it should be possible to move the
script into a separate package to provide this unified interface to all
types of nodes. For now this is considered out of scope.