Fixed Bug OXUIB-27: [L3] Switch between conversation- and no conversation-view loses mail
Root cause: Probably caused by the story OXUI-607. During toggling between normal and thread view the collection gets reset but the complete flag stays on 'true'. So no collection will be loaded as long there aren’t enough mails for pagination in the current folder that triggers the incompleteness.

Solution: Set the complete flag to 'false' manually so that a reload will be triggered.